PrintPool/requirements.txt
brotli==1.0.9
brotlicffi==1.0.9.2
ConfigParser==5.3.0
cryptography==38.0.4
Cython==0.29.32
Django==4.1.4
djangorestframework==3.14.0
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
importlib_metadata==5.2.0
ipaddr==2.2.0
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
mock==4.0.3
mypy_extensions==0.4.3
numpy==1.24.0
ordereddict==1.1
protobuf==4.21.12
pyOpenSSL==22.1.0
python-dotenv==0.21.0
simplejson==3.18.0
typing_extensions==4.4.0
unicodedata2==15.0.0
urllib3_secure_extra==0.1.0
wincertstore==0.2
zipp==3.11.0
